move variational dropout to contrib
